fix(docs): Reordered storybook, added MDX files
