+++ .storybook/main.js
@@ -1,4 +1,9 @@
 module.exports = {
-  stories: ['../src/stories/**/*.stories.js'],
-  addons: ['@storybook/addon-actions', '@storybook/addon-links'],
+  stories: ['../src/stories/**/*.stories.(js|mdx)'],
+  addons: [
+    '@storybook/addon-actions',
+    '@storybook/addon-links',
+    '@storybook/addon-docs',
+  ],
+  showPanel: false,
 }
