+++ package.json
@@ -23,6 +23,7 @@
     "@reach/disclosure": "^0.9.1",
     "@reach/skip-nav": "^0.9.0",
     "@reach/visually-hidden": "^0.9.0",
+    "@storybook/addon-docs": "^5.3.18",
     "@storybook/addon-info": "^5.3.17",
     "emotion-theming": "^10.0.27",
     "eslint-config-airbnb": "^18.1.0",
