+++ src/scss/components/common/lists.scss
@@ -1,0 +1,4 @@
+.list-unstyled {
+  list-style-type: none;
+  margin: 0;
+}
