+++ src/stories/1-docs/0-Introduction.stories.mdx
@@ -1,0 +1,15 @@
+import { Meta } from '@storybook/addon-docs/blocks';
+
+<Meta title="Introduction" />
+
+# Introduction
+
+The COVID Tracking Project  obtains, organizes, and publishes high-quality data required to understand and respond to the COVID-19 outbreak in the United States.
+
+There are a few core tenets that guide our work:
+
+ - **Data quality comes first.** Our engine is the capture and standardization of the variable data that states produce. Our unique cross-functional team does the thing that no one else wants to or can do. It’s grueling, but it generates an extremely valuable dataset. 
+ - We provide an **inclusive, healthy, supportive** work environment for everyone involved
+ - **We operate with maximal openness**, while actively working to make sure our data is used responsibly.
+ - **We maximize value for our network**, not for ourselves.
+ - We support other organizations’ work to provide transparency and accountability for COVID-19 data, regardless of whether they are in industry, government, or media.
