+++ src/stories/2-components/10-Accessibility.stories.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import VisuallyHidden from '@reach/visually-hidden'
 import Alert from '@reach/alert'
-import SkipNavigation from '../components/common/skip-navigation'
+import SkipNavigation from '../../components/common/skip-navigation'
 
 export default {
   title: 'Accessibility utilities',
